♻️ Refactor File to obtain folder features
