+++ lamindb/_file_methods.py
@@ -1,3 +1,4 @@
+from itertools import islice
 from pathlib import Path
 from typing import Optional, Union, overload  # noqa
 
@@ -7,6 +8,7 @@ from lnschema_core.types import DataLike, PathLike
 from upath import UPath
 
 from lamindb._context import context
+from lamindb._file import from_dir, init_file, replace_file
 from lamindb._file_access import filepath_from_file_or_folder
 from lamindb.dev.storage import delete_storage, load_to_memory
 from lamindb.dev.storage._backed_access import (
@@ -172,6 +174,61 @@ def path(self) -> Union[Path, UPath]:
     return filepath_from_file_or_folder(self)
 
 
+# adapted from: https://stackoverflow.com/questions/9727673/list-directory-tree-structure-in-python  # noqa
+def tree(
+    folder: File,
+    *,
+    prefix: str,
+    level: int = -1,
+    limit_to_directories: bool = False,
+    length_limit: int = 1000,
+):
+    """Given a prefix, print a visual tree structure."""
+    if folder.key is None:
+        raise RuntimeError("Virtual folders do not have a tree structure")
+
+    space = "    "
+    branch = "│   "
+    tee = "├── "
+    last = "└── "
+
+    dir_path = UPath(folder.path())
+    files = 0
+    directories = 0
+
+    def inner(dir_path: Path, prefix: str = "", level=-1):
+        nonlocal files, directories
+        if not level:
+            return  # 0, stop iterating
+        # this is needed so that the passed folder is not listed
+        contents = [
+            i
+            for i in dir_path.iterdir()
+            if i.as_posix().rstrip("/") != dir_path.as_posix().rstrip("/")
+        ]
+        if limit_to_directories:
+            contents = [d for d in contents if d.is_dir()]
+        pointers = [tee] * (len(contents) - 1) + [last]
+        for pointer, path in zip(pointers, contents):
+            if path.is_dir():
+                yield prefix + pointer + path.name
+                directories += 1
+                extension = branch if pointer == tee else space
+                yield from inner(path, prefix=prefix + extension, level=level - 1)
+            elif not limit_to_directories:
+                yield prefix + pointer + path.name
+                files += 1
+
+    folder_tree = f"{dir_path.name}"
+    iterator = inner(dir_path, level=level)
+    for line in islice(iterator, length_limit):
+        folder_tree += f"\n{line}"
+    if next(iterator, None):
+        folder_tree += f"... length_limit, {length_limit}, reached, counted:"
+    print(folder_tree)
+    print(f"\n{directories} directories" + (f", {files} files" if files else ""))
+
+
 # likely needs an arg `key`
 def replace(
     file,
@@ -180,8 +237,6 @@ def replace(
     format: Optional[str] = None,
 ) -> None:
     """Replace file content."""
-    from lamindb._file import replace_file
-
     replace_file(file, data, run, format)
 
 
@@ -209,8 +264,6 @@ def __init__(  # type: ignore
     *args,
     **kwargs,
 ):
-    from lamindb._file import init_file
-
     init_file(file, *args, **kwargs)
 
 
@@ -224,3 +277,4 @@ File._save_skip_storage = _save_skip_storage
 File.replace = replace
 File.__init__ = __init__
 File.path = path
+File.from_dir = from_dir
